Add files via upload
This is a newer version of my Bottles of Beer program written at the beginning of the HFJS book.  

I learned how to use objects, methods, this keyword etc, so I thought i'd try it this way.  

It's a little convoluted, but workable.  I'll have to step through the logic at some point.